Add files via upload
These are the Eagle files for the base and module for the L293D IC, the board only requires 2.54 pin headers and a Jumper (or female to female cable). These boards were designed specifically for the Othermill to cut in minutes hence the thick traces and huge spacing between traces. Cutting time is down to 10 minutes on single sided copper clad PCB using the 1/32 bit. Enjoy!